fixed broken animation and effects in forward demo playback after playing a demo backward

+++ codemp/game/bg_pmove.c
@@ -9071,7 +9071,7 @@ CG_SwingAngles
 ==================
 */
 static float BG_SwingAngles( float destination, float swingTolerance, float clampTolerance,
-					float speed, float *angle, qboolean *swinging, int frametime ) {
+					float speed, float *angle, qboolean *swinging, /*int*/float frametime ) {
 	float	swing;
 	float	move;
 	float	scale;
@@ -9158,7 +9158,7 @@ extern qboolean BG_SaberLockBreakAnim( int anim ); //bg_panimate.c
 void BG_G2PlayerAngles(void *ghoul2, int motionBolt, entityState_t *cent, int time, vec3_t cent_lerpOrigin,
 					   vec3_t cent_lerpAngles, vec3_t legs[3], vec3_t legsAngles, qboolean *tYawing,
 					   qboolean *tPitching, qboolean *lYawing, float *tYawAngle, float *tPitchAngle,
-					   float *lYawAngle, int frametime, vec3_t turAngles, vec3_t modelScale, int ciLegs,
+					   float *lYawAngle, /*int*/float frametime, vec3_t turAngles, vec3_t modelScale, int ciLegs,
 					   int ciTorso, int *corrTime, vec3_t lookAngles, vec3_t lastHeadAngles, int lookTime,
 					   entityState_t *emplaced, int *crazySmoothFactor)
 {
